+++ c_src/spectral.h
@@ -8,33 +8,34 @@ struct Grid {
 };
 
 
-double *linspace(float start, float stop, int dim, double factor);
-double *meshgrid(double *linear_grid, int dim, int axis);
+double *linspace(float start, float stop, long long dim, double factor);
+
+double *meshgrid(double *linear_grid, long long dim, int axis);
 
 struct Grid grid(
-    int dim, float pixel_size, float pa, float elong, int elliptic);
+    long long dim, float pixel_size, float pa, float elong, int elliptic);
 
-double *radius(double *xx, double *yy, int dim);
+double *radius(double *xx, double *yy, long long dim);
 
 double *const_temperature(
-    double *radius, float stellar_radius, float stellar_temperature, int dim);
+    double *radius, float stellar_radius, float stellar_temperature, long long dim);
 
 double *temperature_power_law(
-    double *radius, float inner_temp, float inner_radius, float q, int dim);
+    double *radius, float inner_temp, float inner_radius, float q, long long dim);
 
 double *surface_density_profile(
     double *radius, float inner_radius,
-    float inner_sigma, float p, int dim);
+    float inner_sigma, float p, long long dim);
 
 double *azimuthal_modulation(
-    double *xx, double *yy, double a, double phi, int dim);
+    double *xx, double *yy, double a, double phi, long long dim);
 
 double *optical_thickness(
-    double *surface_density_profile, float opacity, int dim);
+    double *surface_density_profile, float opacity, long long dim);
 
 double bb(double temperature, double wavelength);
 
 double *intensity(
-    double *temperature_profile, double wavelength, double pixel_size, int dim);
+    double *temperature_profile, double wavelength, double pixel_size, long long dim);
 
 #endif
